Rename: View/ViewFactory --> Factories/ViewFactory
