+++ src/Decorators/Fields/Textarea.php
@@ -18,7 +18,7 @@ declare(strict_types=1);
 
 namespace TypistTech\WPBetterSettings\Decorators\Fields;
 
-use TypistTech\WPBetterSettings\Views\ViewFactory;
+use TypistTech\WPBetterSettings\Factories\ViewFactory;
 use TypistTech\WPBetterSettings\Views\ViewInterface;
 
 /**
